refactor: 메시지 말풍선 렌더링 공통 컴포넌트 분리
- ChatMessageList 내부의 메시지 말풍선 렌더링 로직을 MessageBubbleList로 분리
- 발신자명, 프로필, 시간, 말풍선 UI를 공통 컴포넌트에서 처리하도록 수정
- 읽지 않음 표시는 showUnreadMarker 옵션으로 제어하도록 수정
